Moved security related files to own folder
